Add rmrk core event handling (#7)
