+++ schema.graphql
@@ -201,7 +201,6 @@ type Battlepass @entity {
     price: BigInt!
     createdAtBlock: Int!
     updatedAtBlock: Int!
-    collectionId: Int!
 
     # Metadata:
     cid: String!
@@ -211,4 +210,23 @@ type BattlepassNft @entity {
     id: ID!
     owner: Identity!
     battlepass: Battlepass!
+    nft: Nft!
+}
+
+type Nft @entity {
+    id: ID!
+    owner: String!
+    metadata: String!
+    equipped: Boolean
+    pending: Boolean!
+    transferable: Boolean!
+    collection: NftCollection!
+}
+
+type NftCollection @entity {
+    id: ID!
+    issuer: Identity!
+    metadata: String!
+    max: Int
+    symbol: String!
 }
