+++ src/balances.ts
@@ -19,7 +19,7 @@ import { Event } from './types/generated/support'
 import { Context } from './processor'
 import { HistoricalBalance, AccountBalance } from './model'
 import { encodeId } from './common/tools'
-import { upsertIdentity } from './mappings/util/db/identity'
+import { upsertIdentity } from './common/db/identity'
 
 
 export async function saveBalancesAccounts(ctx: Context, block: SubstrateBlock, accountIdsHex: Set<string>) {
