+++ src/mappings/battlepass/index.ts
@@ -1,15 +1,15 @@
-import { Pallet } from '../types/pallet'
+import { Pallet } from '..'
 import { handleBattlepassCreatedEvent } from './events/battlepassCreated'
 import { handleBattlepassClaimedEvent } from './events/battlepassClaimed'
 import { handleBattlepassEvent } from './events/battlepass'
 
 export default {
-	name: 'Battlepass',
-	callHandlers: {},
-	eventHandlers: {
-		'BattlepassCreated': handleBattlepassCreatedEvent,
-		'BattlepassClaimed': handleBattlepassClaimedEvent,
-		'BattlepassActivated': handleBattlepassEvent,
-		'BattlepassEnded': handleBattlepassEvent,
-	},
+    name: 'Battlepass',
+    callHandlers: {},
+    eventHandlers: {
+        'BattlepassCreated': handleBattlepassCreatedEvent,
+        'BattlepassClaimed': handleBattlepassClaimedEvent,
+        'BattlepassActivated': handleBattlepassEvent,
+        'BattlepassEnded': handleBattlepassEvent,
+    },
 } as Pallet
