+++ src/mappings/control/events/memberUpdated.ts
@@ -2,25 +2,25 @@ import { Context, Block } from '../../../processor'
 import { Event } from '../../../types/generated/support'
 
 import { getMemberUpdatedData } from './getters'
-import { getOrgMember } from '../../util/db/getters'
+import { getOrgMember } from '../../../common/db/getters'
 
-import { arrayToHexString, addressCodec } from '../../util/helpers'
+import { arrayToHexString, addressCodec } from '../../../common/tools'
 import { ObjectNotExistsWarn } from '../../../common/errors'
 
 
 async function handleMemberUpdatedEvent(ctx: Context, block: Block, event: Event, name: string) {
-	const [ memberAddress, memberOrgId, state ] = getMemberUpdatedData(ctx, event)
-	let address = addressCodec.encode(memberAddress)
-	let orgId = arrayToHexString(memberOrgId)
+    const [ memberAddress, memberOrgId, state ] = getMemberUpdatedData(ctx, event)
+    let address = addressCodec.encode(memberAddress)
+    let orgId = arrayToHexString(memberOrgId)
 
-	let member = await getOrgMember(ctx.store, orgId, address)
-	if (!member) {
-		ctx.log.warn(ObjectNotExistsWarn(name, 'Member', `${orgId}-${address}`.toLowerCase()))
-		return
-	}
-	member.state = state.__kind
+    let member = await getOrgMember(ctx.store, orgId, address)
+    if (!member) {
+        ctx.log.warn(ObjectNotExistsWarn(name, 'Member', `${orgId}-${address}`.toLowerCase()))
+        return
+    }
+    member.state = state.__kind
 
-	await ctx.store.save(member)
+    await ctx.store.save(member)
 }
 
 export { handleMemberUpdatedEvent }
