+++ src/mappings/control/index.ts
@@ -1,4 +1,4 @@
-import { Pallet } from '../types/pallet'
+import { Pallet } from '..'
 import { handleMemberAddedEvent } from './events/memberAdded'
 import { handleOrgCreatedEvent } from './events/orgCreated'
 import { handleMemberRemovedEvent } from './events/memberRemoved'
@@ -7,13 +7,13 @@ import { handleOrgUpdatedEvent } from './events/orgUpdated'
 
 
 export default {
-	name: 'Control',
-	callHandlers: {},
-	eventHandlers: {
-		'OrgCreated': handleOrgCreatedEvent,
-		'OrgUpdated': handleOrgUpdatedEvent,
-		'MemberAdded': handleMemberAddedEvent,
-		'MemberRemoved': handleMemberRemovedEvent,
-		'MemberUpdated': handleMemberUpdatedEvent,
-	},
+    name: 'Control',
+    callHandlers: {},
+    eventHandlers: {
+        'OrgCreated': handleOrgCreatedEvent,
+        'OrgUpdated': handleOrgUpdatedEvent,
+        'MemberAdded': handleMemberAddedEvent,
+        'MemberRemoved': handleMemberRemovedEvent,
+        'MemberUpdated': handleMemberUpdatedEvent,
+    },
 } as Pallet
