+++ src/mappings/rmrkCore/index.ts
@@ -1,0 +1,12 @@
+import { Pallet } from '..'
+import { handleCollectionCreatedEvent } from './events/collectionCreated'
+import { handleNftMintedEvent } from './events/nftMinted'
+
+export default {
+    name: 'RmrkCore',
+    callHandlers: {},
+    eventHandlers: {
+        'CollectionCreated': handleCollectionCreatedEvent,
+        'NftMinted': handleNftMintedEvent,
+    },
+} as Pallet
