+++ src/mappings/sense/index.ts
@@ -1,12 +1,12 @@
-import { Pallet } from '../types/pallet';
-import { handleEntityCreatedEvent } from './events/entityCreated';
-import { handlePropertyUpdatedEvent } from './events/propertyUpdated';
+import { Pallet } from '..'
+import { handleEntityCreatedEvent } from './events/entityCreated'
+import { handlePropertyUpdatedEvent } from './events/propertyUpdated'
 
 export default {
-	name: 'Sense',
-	callHandlers: {},
-	eventHandlers: {
-		'EntityCreated': handleEntityCreatedEvent,
-		'PropertyUpdated': handlePropertyUpdatedEvent,
-	}
-} as Pallet;
+    name: 'Sense',
+    callHandlers: {},
+    eventHandlers: {
+        'EntityCreated': handleEntityCreatedEvent,
+        'PropertyUpdated': handlePropertyUpdatedEvent,
+    }
+} as Pallet
