--- src/mappings/types/identity.ts
@@ -1,12 +1,0 @@
-export interface IdentityUpsertData {
-	address: string;
-	displayName: string | null;
-	legalName: string | null;
-	email: string | null;
-	riot: string | null;
-	image: string | null;
-	twitter: string | null;
-	web: string | null;
-	web3name?: string | null;
-	discord?: string | null;
-}
