+++ src/storage/index.ts
@@ -3,11 +3,13 @@ import * as flow from './flow'
 import * as signal from './signal'
 import * as sense from './sense'
 import * as battlepass from './battlepass'
+import * as rmrkCore from './rmrkCore'
 
 export const storage = {
     control,
     flow,
     signal,
     sense,
-    battlepass
+    battlepass,
+    rmrkCore
 }
